Use Preferences for setting Log Level
Summary:
Uses the `MobileDevice.framework` initialization routine in order to set the log level. This will mean that the `asl` logger is setup appropriately. Unfortunately `asl_add_log_file` doesn't behave as expected on some more recent versions of macOS, so logging to a file may not work.

We will at least get more verbose logging on `stderr` where appropriate, which is great.

Reviewed By: dtorres

Differential Revision: D6660212

fbshipit-source-id: 5ad230ed2faaca5bde30eeb87ca676dbf58d515b