Make FBSimulatorBridge Creation Asynchronous
Summary:
There is a hard-dependency on spawning the `SimulatorBridge` Agent, which now means we would have to `await:` the Agent Operation Future. Instead of doing this we can propogate-out this asynchrony. This also makes a lot of the error handling much easier for us.

During this, I also found that there are process leaks if the Framework is repeatedly invoked from separate host processes. A `SimulatorBridge` will be spawned *per host process* as they are uniqued on this pid. Instead of doing this we can share the same `SimulatorBridge` binary and bind to it. We can also optimize the "slow path" by running the poll-loop and the launch of the `SimulatorBridge` process in parallel.

Reviewed By: marekcirkos

Differential Revision: D6691999

fbshipit-source-id: 1b76603cf1108f6b1a33698566f2346572b696c6