Use FBProcessOutput in FBTask
Summary:
Removes the Private `FBTaskOutput` class cluster inside `FBTask` and replaces it with usage of `FBProcessOutput`. This means that all of the work is `FBProcessOutput` is now consolidated and is asynchronous.

The `FBProcessOutput` implementations can then be re-used between Agent Launches and Tasks, which will allow for attaching all kinds of other consumers of the `stdout` and `stderr` of an agent, including using loggers or pushing into a file or buffer.

Further to this, the `FBProcessOutput` can then be extended to expose the contents as a file path that's also a fifo. This is useful for both the shimming of logic tests and for the reading of output from an App Launch. The latter is not yet supported but will enable reading the output of a launched application process in the same way that we can do for launching an Agent.

Reviewed By: marekcirkos

Differential Revision: D6682562

fbshipit-source-id: 407c2bd60c3f81ccdd95a1bf84d228aedb2cdff2