All outcomes of FBFuture should evict cancellation handling
Summary: Currently, it is possible to pass in a block for `respondToCancellation` that is not nullified until the `FBFuture` instance deallocates. This makes it easier to create a retain cycle than should be necessary. Therefore we should always evict the cancellation handler whenever the future resolves. Since `resolveAsCancelled` is a private method, we can use this to pop the cancellation and return this to the caller. I also noticed that there was double calling of `resolveAsCancelled` but this should be harmless.

Reviewed By: dtorres

Differential Revision: D6723508

fbshipit-source-id: 53550943dd50e98ca2d03a786b3c174d3c5fff55