Add Task from stdin data.
Summary: Adds the ability to pass in some `NSData` to `stdin` directly rather than having to pump it manually from above. This makes the data passing asynchronous so the caller does not have to worry about blocking the `write` whilst pushing the data through the task. For processes that stay alive for the duration of their input, this is desirable as the process will terminate when the input ends and all the `NSData` has been written

Reviewed By: asm89

Differential Revision: D6772052

fbshipit-source-id: 1a5c0a1ae1dad53bc916e98238e8a55420eefa13