Make -[FBFileReader fileReaderAtPath:] non-blocking
Summary:
The current implementation of this, is such that the caller will be blocked on an `open` in the calling thread, which cuts against the whole point of this being a non-blocking API.

Instead of just passing a path and assuming that this exists at the time that the file is open, we can instead build this into an async constructor without changing any internal semantics. The problem with hiding this in the `startReading` call is that in the case of a `fifo` the file may not yet be open for writing. Since `dispatch_io` will open the file in a non-blocking manner, we'll ready an empty file that isn't open yet.

Additionally, using the `-[NSFileHandle fileHandleForReadingAtPath:]` will *assert* in the event that the `fifo` could not be opened because the it has never been `open`ed at the other end. It does fail if the path doesn't exist but the assertion here is really not helpful as it will blow up the test execution in the event that something goes wrong on the launch of the process, killing a huge amount of diagnositcs. By using `open` directly, we can get the `errno` and report this to users.

Reviewed By: marekcirkos

Differential Revision: D6757742

fbshipit-source-id: a0f082cbde07da1f66349e391a8e7a105aeaddab