Orthogonalize FBXCTestConfiguration & FBXCTestDestination
Summary: In order to run a test via the CLI we have a 'destination' that describes how to obtain a target. The runner classes take a target, so passing a target to a runner requires creating a destination object unecessarily. This makes the Framework more flexible and pushes `fbxctest` only classes into the correct place

Reviewed By: marekcirkos

Differential Revision: D6847573

fbshipit-source-id: 5f5f304d9ad9d854927b1151cd533bc2f023e74b