Ensure that "error" is handled the same as "failure"
Summary:
The `error` event is not properly handled and represents a caught exception, rather than an assertion failure within the test case. Whatever the intracacies of this, the "error" case clearly needs to be handled appropriately.

I've broken down into some more discrete methods to make this a little easier to follow.

Reviewed By: dtorres

Differential Revision: D6913711

fbshipit-source-id: 2f7c62d8553e7368e3e17b984665d78269e57dd2