Make FBSimulatorAgentOperation only fetch process-info optionally
Summary:
For short lived processes (i.e. launchctl list) it is possible for a process to be removed from the process info table before we can fetch this info.

For this reason, the availability of this Process Information on an Agent Operation is now optional. We will always have the `pid` which we have exposed already. We can attempt to fetch this in the background and this will fullfill the slot, so that the Future is a serialized container of this value.

I've removed the de-duplication of operations as well. This is unecessary now since we can guarantee de-duping.

Reviewed By: dtorres

Differential Revision: D6998394

fbshipit-source-id: 1386afa1b31077353dde09253646999b2e0131f4