Don't wait for Framebuffer Teardown Synchronously
Summary:
The current implementation of `-[FBSimulatorConnection terminateWithTimeout:]` will block on the `dispatch_group` Synchronously. This is a poor fit with the rest of the Framework and means that the main queue is blocked just for the purpose of waiting on some condition.

Instead we can move the timeout behaviour outside and use the asynchronous variants.

Reviewed By: marekcirkos

Differential Revision: D6922654

fbshipit-source-id: 9d03b31a994c2ad6cdb50e16026f12db03e5a32b