Assert that chained/fmapped futures are Futures
Summary: It's easy to shoot yourself in the foot by returning a non-future value from the blocks of these calls. The compiler goes some way to make this less of a problem, but the inference is never perfect. For this reason, catching these calls before getting an unrecognized selector is going to help out a little.

Reviewed By: dtorres

Differential Revision: D7009155

fbshipit-source-id: 461b84156da90a8535a0f8f1e759c732dd2d9449