Write eof in shim fifo
Summary:
In the teardown of a test listing we currently close the file when the process has terminated. To ensure all of the data has been written we also want to ensure that the file has been fully exhausted, there is a race condition that can occur between being signalled that the process has exited and the full draining of the fifo.

We can instead wait for and end-of-file and then close the reader. To do this we have to close the file handle on the writing end (the shim). The current implementation causes the reader to become "cancelled" which can race with the draining.

The future will ensure that the callback is not called twice. With use of a fifo that sends an `eof` normally, I've found that the `eof` may never be sent to consumers. Fixing this will mean that we can generalize a `fifo`.

Reviewed By: phyllipecesar

Differential Revision: D7081674

fbshipit-source-id: 97c077ad619df97cc192e2124cef7647cc4033cf