Resolve Completion in dispatch_io_read as well as constructor
Summary:
The future will ensure that the callback is not called twice.

If the io channel is closed before reading has finished we will also need to handle the `ECANCELLED` case so that stopping reading does not result in the future resolving in error. From the `dispatch_io_close` documentation:

> If the DISPATCH_IO_STOP option is specified in the flags parameter, the system attempts to interrupt any outstanding read and write operations on the I/O channel. Even if you specify this flag, the corresponding handlers may be invoked with partial results. In addition, the final invocation of the handler is passed the ECANCELED error code to indicate that the operation was interrupted. If you do not specify the DISPATCH_IO_STOP flag, read and write operations on the channel run to completion as normal.

Reviewed By: dtorres

Differential Revision: D7040520

fbshipit-source-id: f7ad903ae14942866e39fc2f11f31cd1eb33f263