Translate FBProcessOutput to a fifo
Summary:
Allows conversion of an `FBProcessOutput` to be mounted to a file path. This will allow direct usage of `FBProcessOutput` in other contexts where we wish to consume output as it is written to a file. Some `CoreSimulator` APIs as well as the shim writing in `fbxctest` *must* write to a file instead of a pipe/filedescriptor. Bridging in this way means that we can have full equivalence.

For now this doesn't enable any new features, but it will mean that we can perform in-process consumption of the output of a launched Application (CoreSimulator can only write to a file) which is going to be very useful for improving the output of Application Tests.

Placing this into a private class cluster also makes implementations for some cases very simple. For a `FBProcessOutput` that writes to a file, we can just pass down this file. This is also true of `/dev/null`.

Reviewed By: dtorres

Differential Revision: D7040533

fbshipit-source-id: 64e881f4fae723a4822269ac052e58b884b48a28