In-Process Output from an Application launch.
Summary:
Up until now, it hasn't been possible to consume the output of an Application process in-process. Now we have an API for yeilding a file path from an arbitrary `FBProcessOutput` instance, we can re-direct output via a `fifo`.

This also has the nice effect that we can delete a lot of the cludge around diagnostics and this being an implicit dependency. Workflows that launch Applications (i.e. Application Tests) will also be able to use this in-process output for additional diagnostics during the execution of the test run.

One important part here is passing the `stdout` and `stderr` objects into the `FBSimulatorApplicationOperation` so that they can be retained for at-least the lifecycle of the Application. Without this, there may be nothing to retain the reader/writer pairings.

Reviewed By: dtorres

Differential Revision: D7040547

fbshipit-source-id: e1ae122d25d30391dd86552bcae29376529073fe