Add signalling to FBTask
Summary:
We haven't previously surfaced the ability to send signals other than `SIGTERM` to the wrapped task. This adds this behaviour and corresponding tests.

Since we're also exposing this in a public API, there needs to be additional resilience in how we perform the teardown. This uses dispatch for synchronization as well as some future states to prevent async teardown from executing twice

Reviewed By: dtorres

Differential Revision: D7183319

fbshipit-source-id: 9a51cc6bfa631ace9aa267b0afdb41138566bdd5