Support list tests of an app test
Summary:
Previously fbxctest cannot list tests of an app test bundle because the test binary contains unresolved
symbols which should be exported in main executable. But since we are using "xctest" executable to load the
test bundle, those symbols cannot be resolved. (more details in T24151331)

This can by fixed easily by using the target app as the executable to load the test bundle and then run the
shim code to dump the test lists. Since we're running the app directly without installation, An additional step
required is to copy XCTest.framework to app's rpath so it can be found by dyld.

Reviewed By: lawrencelomax

Differential Revision: D7166532

fbshipit-source-id: 2dcbb00acc750aa5fb68876a920d0a468c1bee83