Fix un-needed closing of reader on eof
Summary:
The implementation of the `FBProcessInput` erroneously calls `detach`, which means that the following will occur:
1) A Task is created that has the stdin connected (i.e. a consumer)
2) An eof is send by the client after all data is sent.
3) The `FBProcessInput` class sends an eof to the `FBFileWriter` which is async
4) `detach` is called, tearing down the writer
5) The data is never written and the pipe breaks
6) The subprocess terminates on broken pipe

This is seen in the `testInputReading` test, as `cat` terminates before it is sent the data. The writer is (deliberately) asynchronous to avoid blocking the caller, but internally we must account for this.

Instead, we will let the owner of the `FBProcessInput` class do the heavy lifting in terms of co-ordinating the detachment.

Reviewed By: marekcirkos

Differential Revision: D7210318

fbshipit-source-id: b199c3da4fa457292839fcab7edf24454e3d6838