Generics for FBTask
Summary:
Adds Generic annotations to `FBTask` in order that the stdin/stdout/stderr can be (optionally) strict. There's no enforcement of this at runtime, but code-comlpletion an inference should make it a little harder to shoot yourself in the foot.

This also makes the types of the constructor very obvious. This has also been documented.

Reviewed By: marekcirkos

Differential Revision: D7210386

fbshipit-source-id: a8ac3789c3ce79e3a965c46774cfee9a5c3ef690