Loggers should handle thread-safety
Summary:
The fact that a logger must have it's calling-queue specified makes them much harder to use. Instead of doing this, we can just push this down into the implementation to handle. If it is called from multiple queues then it's the job of the implementation to do the right thing. `os_log` is documented as being thread safe, so we don't have to worry about the system logger.

In fact there's only one instance here that we need to care about, and that's the consumer based logger. This is a synchronous writer so writes need to be serialized. The critical section here is placed into a `synchronized` section on the *writer* so that child loggers are all synchronized. This could eventually be made an async queue based writer, but for the time being we should have the behaviour that the queue is flushed before returning to the caller.

Reviewed By: marekcirkos

Differential Revision: D7513516

fbshipit-source-id: 239dbcfd70127919b6820a33026d6553d641ee2e