Add Interface for obtaining crash reports
Summary:
Adds an implementation of a push-based way of determining whether a process has crashed.

This ends up using the `FSEvent` API for ingesting new paths that have been created. We can then associate a crash with the pending notification and dispatch this. This also saves polling on the `DiagnosticReports` path and crash reports are globally de-duplicated, so we don't end up ingesting the same crash report twice. Being notification-based may also decrease the latency to get a crash report. From some simple usage, I've found that this does result in the Crash Integration tests finishing faster.

Reviewed By: marekcirkos

Differential Revision: D7192713

fbshipit-source-id: d5a7153091bb15cf0e036f61cfaff9356bcba710