Name Simulator & Device specific loggers
Summary: The naming of loggers is ad-hoc and introduces unecessary `:`'s for formatting. With `os_log` these are put in square braces. If we always pass down from above we'll get the prefixed-logging for domains that don't add their own name. This makes it far easier to disambiguate which Simulator is responsible for a given log message.

Reviewed By: marekcirkos

Differential Revision: D7652237

fbshipit-source-id: c7792566779b1bdc312b28c87c240b9272a8f530