Add Detection of Boot Status Stalls
Summary: Adds the ability to detect a "stall" of a boot by checking the last time we have a new object for Boot Info. These are recieved as XPC Notifications in `SimDevice` and consequently set on the property. It is possible to use KVO here, but just augmenting the existing logging for the time being.

Reviewed By: javache

Differential Revision: D7666634

fbshipit-source-id: b8395c96c923645b9cf2fa8697886181a59160a6