Timeout Simulator Fetch if > 80% of total timeout
Summary:
In rdar://35628711 (macOS 10.13.4), Simulator booting has regressed substantially. Since Simulator booting is a large part of the test process, it's important that we have tests that obey the timeouts. In this case we need to be sure that a Simulator boots within a reasonable proportion of the timeout. We should therefore place a timeout on section of the test execution. The test will fail and obey the timeout, rather then getting stuck until it does.

This also exposes that timeouts are inconsistently applied. For this reason, I've now moved the timeout down towards the base runner, which can apply timeouts consistently.

Reviewed By: skotchvail

Differential Revision: D7756931

fbshipit-source-id: c11f657a876bc5d42daf4060e8cbed2059e71a91