Add Timeout for _XCT_testBundleReadyWithProtocolVersion: to be called
Summary: If the Application that hosts the Test Bundle (often a `RunnerApp`) crashes or otherwise fails, the Future that resolves upon connection to the Test Manager will never resolve. In order to fix this, we can place a timeout on this future being resolved. 20 Seconds is a pretty reasonable value here, as this should resolve *very fast*. The Test Host Application will have been launched at this state, so the Simulator will be in a booted state.

Reviewed By: marekcirkos

Differential Revision: D7788386

fbshipit-source-id: 5222990e7579a084445ac77598a9406fc98dd431