Expose "Name" of a crash log in FBCrashLogInfo
Summary: This is the "Path-Unique" name that identifies a Crash Log, so that it can be re-called later on. I'm exposing this on the `FBCrashLogInfo` class to make it clear that this is different from the "Identifier" which is more analagous to the the Bundle ID of a process. By placing this on `FBCrashLogInfo` it makes this class aware of it's own "key" which helps for re-call

Reviewed By: marekcirkos

Differential Revision: D8215417

fbshipit-source-id: 119cbcc8649e55a294045d3204d285e64c9ffae9