Allow asynchronous usage of a AMDevice Connection
Summary:
The gnarlyness of `futureForDeviceOperation:` means that it's very tricky to be explicit about the lifecycle of a connection. For this reason, I've moved out ownership into a separate component that can track (and log) this.

With this in mind we can then improve the behaviour to disconnect from the device when any chaining has completed.

Differential Revision: D8258180

fbshipit-source-id: 7e253bd4f49f61f228d0375e7a9bb11f3a7d2f61