Add FBFutureContext
Summary:
Adds a new wrapper around `FBFuture` that allows for context to be torn down.

This is very useful in APIs that return a connection for consumers to use. The connection may eventually need to be torn down, so this knowledge can be hidden from consumers. A regular future can then be extracted in the normal way, with precise semantics for the lifcycle of the context.

Reviewed By: rlopezdiez

Differential Revision: D8304732

fbshipit-source-id: f96d983ca795b25cab06c678a46ef01bfade90ad