Use FBFutureContext in FBAMDevice
Summary:
With the move to more Asynchronous APIs in `FBDeviceControl`, Future-based methods for obtaining a "connected" device, as well as other connections to various channels such as a Service Connection or AFC were used.

However, given that the consuming of these connections is now Asynchronous, we need to be much more precise about the lifecycles of these connections so that we don't leak them or close them before they are used.

In order to do this, we can use the new `FBFutureContext` APIs. This allows for consumers to define how long they will use these connections. The internals of the cleanup don't need to be exposed, which makes it impossible for consumers to forget to perform some cleanup.

Reviewed By: zsol

Differential Revision: D8305670

fbshipit-source-id: 82e18864a45f5d51880004ccabc6f3bc6e134110