Prevent Concurrent Access of AMDevice
Summary:
If two consumers utilize a device at the same time, they can cause multiple connections and therefore failure.

In order to prevent this, we have a per-device notion of an internal 'USE' queue that can also hold a future for the next consumer that comes along. This is an effective way of synchronizing all consumers.

This also then enables using the same connection over multiple usages.

Reviewed By: zsol

Differential Revision: D8560829

fbshipit-source-id: 13b89d0b27091efb86c0d87219cef75e7ea5bbad