Add FBAMDeviceServiceManager
Summary:
Adds a way of de-duplicating and re-using existing connections in a way that is completely transparent to the consumer. This currently de-dupes the house_arrest service.

This works by memoizing a context manager of a connection. The semantics are identical as they are for utilizing a device, except that we can control the lifecycle.

Reviewed By: marekcirkos

Differential Revision: D8661600

fbshipit-source-id: 39e62cc948dbeb5cd795ea6d12d014d0b37b2146