Remove diagnosticAvailable:
Summary:
In an effort to consolidate Simulators & Devices, this stuck out as a useless artifact. It also goes against having an API where writing to a file is explicit.

These are only used for screenshots and video recording. However, each of these now take an output path as an argument in the "Commands" protocols, so those should just be used instead.

Reviewed By: marekcirkos

Differential Revision: D8799542

fbshipit-source-id: 674706da2ab399326137a7e69f88f570e4e5f8c5