Remove FBFramebufferFrameGenerator
Summary: This is now no-longer needed and we can simplify a lot of the code by removing it. Now, all usages of a Simulator's Framebuffer should go through the usage of attaching to an `IOSurfaceRef` and it's damage rectangles. This is substantially more efficient and means we don't have to re-invent the wheel.

Reviewed By: marekcirkos

Differential Revision: D8800033

fbshipit-source-id: 94a672f24fb9ba739d59709ff10d888d4609074a