Sharable Context
Summary:
This adds the ability to share a context in the `FBFutureContextManager`. For some cases, it can be beneficial to share a context between interested parties. For example, connecting to an `AMDevice` instance can easily be shared. In other cases, such as an AFCConnection this may make less sense.

All existing consumers have been set to `NO` for enabling sharing. In future changes we can opt these components in to sharing.

Differential Revision: D8891920

fbshipit-source-id: df50a52a431f088d62f35ba2250c3eefdc8918f1