Opt-In FBAMDevice to Context Sharing
Summary:
This will allow multiple consumers to re-use the same context. In this case the "context" is the connection to a device.

The net win here is that two requests that use different AMDevice APIs can run asynchronously with each other. A request to start house arrest and a request to install an app will be serialized on queues relative to the AMDevice, but there's no need for repeated connections.

Reviewed By: dguerri

Differential Revision: D8891919

fbshipit-source-id: 822753f81a0ba0c9da2643b36f7ee2968a729ca7