Back out "[FBDeviceControl] Don't wait to teardown device / service connections"
Summary:
This was causing a spike in "Failed to start house_arrest" errors when we perform a lot of file operations in quick succession.

This was originally a workaround for iPhone 6 / iOS 10 install and is not needed now as the previous diff in the stack fixes install in a better way.

Reviewed By: marekcirkos

Differential Revision: D9545011

fbshipit-source-id: 9ba6c1fb3eae2d6cd4dce743dd5af8e5c68a880f