Add -[FBCrashLogCommands pruneCrashes:]
Summary:
Instead of exposing the details of the Crash Log Store, it makes some sense to be able to control the availability of crash logs for a given target.

For Simulators this now means we have to specialize since we need to have knowledge of the UDID being present in (some) logs but not others.
For Devices we need to be able to prune crashes by talking to the device

Reviewed By: zeyadsalloum

Differential Revision: D8595775

fbshipit-source-id: 5fc559191d881c591842c425e1dd026f51f78cbd