crash when connecting a device
Summary:
the notifier was crashing when a device connects. because the amdevice was set after the notification fires and so the notifier was trying to read the AMDevice's name, state etc when they were not there yet.

this was casuing list-targets to not report devices correctly and also was causing connect UDID to not work with devices

Reviewed By: c-ryan747

Differential Revision: D13194864

fbshipit-source-id: 838abedf877982bf6571650b29524c8995d8ee5b