added a logger to FBFileReader
Summary:
Not calling stopReading causes a crash in libdispatch. so it's very important to detach from an FBProcessOutput before it gets deallocated.

I added a logger so we can log an error when it gets deallocated before all file handlers were detached so that it's easier for this problem to be detected sooner

Reviewed By: lawrencelomax

Differential Revision: D13254727

fbshipit-source-id: c0130195fe5a78162f675c2ef068e6cc91356369