Write Unsupported Device Log Arguments message back
Summary: This is currently only shown in the internal logs, so consumers of the message might not be aware that the arguments are unsupported. rdar://38452839 was made and has been duped to use the `log` cli instead of this MobileDevice API. When Apple fixes this we'll be able to change the implementation over to use the `log` cli instead, with all of the relevant arguments supported.

Reviewed By: zeyadsalloum

Differential Revision: D13503904

fbshipit-source-id: 10004e8fd3c1cf2af80e4c9f1578342f1ff8d717