Fix File Handle Teardown Semantics
Summary:
There's some more issues with teardown sematics that means that consumers of file handles can wait longer than they need to, to teardown a pipe.

Now the case analysis has been consolidated and it's possible to disambiguate between cases. The primary cause of the problem is that:
- If a file read finishes successfully, we don't call `dispatch_io_stop`. You should *always* call this, regardless of whether we are cancelling mid-flight or not.
- The futures returned to clients should only care about the sending of an eof. They don't need to care about the internal details of any asynchronous channel teardown. This means we can return earlier.
- Since we now have the self-capture in order to keep the IO Channel and FD alive during a read, we don't necessitate `stopReading` to be called. If anyone does call this after a file read has finished, we can just short-circuit immediately

The reason for all these changes is that in the successful-state, we didn't end up calling `dispatch_io_stop` which is *required* in order for the futures to evaluate. Now we do this in the right place and only use `DISPATCH_IO_STOP` in the cancellation case.

Reviewed By: c-ryan747

Differential Revision: D13537802

fbshipit-source-id: bd30bef8ed0d0fb79cae1bebfb14a33bd80bd748