Add -[FBDevice pathForDeviceSupportDiskImageWithError:]
Summary:
In order for debugging to work, we need the developer disk image for a given path. The most appropriate way of doing this is to search through known paths that contain the disk image, and then find the most relevant build that matches. There are not disk images for patch releases, but there are for minor versions. To get this working this means that we need to match the build number explicitly (i.e. `16B91`) or else match on the major & minor versions.

This means we'll then select the appropriate disk image to mount on the device.

Reviewed By: zeyadsalloum

Differential Revision: D13572860

fbshipit-source-id: 7d7e76450f4f254d4ff11231da44a1887e8723a0