Add FBDispatchDataConsumer
Summary:
Over time we can eventually migrate more and more callsites to use dispatch_data, which is of great help when dealing with pairs of sockets. I have some new scenarios coming that just shuttle dispatch_data around, so exposing APIs that make this a possibility makes the bridging easier in thise case.

This requires no changes of existing API users since the methods behave as before, they just use a wrapping class that does the conversion forward and back.

Reviewed By: zeyadsalloum

Differential Revision: D13596759

fbshipit-source-id: 417341bdb2b113a231d8fa31d71e811dbd9e31e3