Prevent race condition in data delivered after dispatch_io_stop
Summary: There is a (rare) race condition that means a consumer of data can recieve and eof followed by data. This is clearly erroneous and exists because of the behaviour of `dispatch_io_stop`. It doesn't guarantee that no more data will be supplied to the block. As such we have to ensure this is only ever sent once when the IO Channel is torn down.

Reviewed By: zeyadsalloum

Differential Revision: D13634937

fbshipit-source-id: 9a79505c467528adaf40131cafa516cdf257b185