Fix teardown semantics of FBFileWriter
Summary: There are some problems in the way that writing is handled in `FBFileWriter` in the async case that means that we can end up in a scenario where we are closing the file descriptor behind the back of the IO channel. This fixes this by ensuring that the IO channel is fully torn down before we do this.

Reviewed By: zeyadsalloum

Differential Revision: D13636345

fbshipit-source-id: d76ba5dcc704e006d6c6d003329a9d54bdbf1258