Fix Cancellation of finishedReading
Summary:
The `finishedReading` call aliases the `ioChannelFinishedReadOperation` future. If this future is cancelled, then the cancellation responder will also end in cancellation, which is an error condition.

What should happen is that the cancel responder should finish when the `ioChannelFinishedReadOperation` resolves, but avoid cancelling the `ioChannelFinishedReadOperation`. To do this we can return a new future that is resolved from the `ioChannelFinishedReadOperation` but not cancelled by it

Reviewed By: c-ryan747

Differential Revision: D13653140

fbshipit-source-id: ec94f839c5f3def701f85a51e776ad3f86e0158a