Ensure that consumeEndOfFile is sent when reading is done rather than channel closed
Summary:
There's a subtle bug that appears in how channel closing occurs. Ultimately, `dispatch_io_read`'s 'done' is the ultimate arbiter of if we've reached the end of a file. This is distinct from the `dispatch_io_create` callback that deals with the control of the channel rather than the read operation itself.

This poses a problem for `FBLogicTestRunStrategy` which will wait for an eof indefinately even if the reader has stopped as `consumeEndOfFile` is not sent to the consumer. Now we can ensure that this doesn't happen as the `done` in `dispatch_io_read` signals this instead.

Reviewed By: c-ryan747

Differential Revision: D13653139

fbshipit-source-id: 59fa25d899d863a978c9b95483039c675c72507c