Add FBDispatchDataConsumer for FBFileWriter
Summary:
There were racey deallocations in D13596759, so this fixes issues related to the excessive copying and wrapping of mutable data from `NSData`. This incorporates all of the necessary fixes and only changes `FBFileReader`.

The root cause was that the IO Channel was disposed behind the back of data that was being pushed into it, as a result the data was over-released before it had been flushed.

Reviewed By: c-ryan747

Differential Revision: D13636425

fbshipit-source-id: b2a82a8456d5cc305c641e7a8765885a9c49160e