Socket-only Transport for testmanagerd
Summary:
Instead of having the implementors wrap a `DTXTransport` we can perform the translation inside `XCTestBootstrap` itself. This removes this depdendency from each of the device/simulator frameworks.

Since each of the implementors allocates the socket, it's also important that it knows how to tear it down. This is exactly the use-case for `FBFutureContextManager`. Each of the implementations can provide their own socket and tear it down in the same place.

Reviewed By: zeyadsalloum

Differential Revision: D14151895

fbshipit-source-id: cbc8c9847b60615fa17b87bd33ec56b56392fd52