Remove FBiOSDeviceOperator
Summary:
We're finally at the point that we can kill the "Device Operator" for good. All of the relevant commands have been distributed amongst the command classes.

The bigger impact is that we can get rid of the Xcode classses that are used in `FBDeviceControl` and move further in the direction of doing everything through `MobileDevice.framework`. This removes a huge amount of fragility from the system, and simplifies linking substantially.

Reviewed By: zeyadsalloum

Differential Revision: D14163721

fbshipit-source-id: 7c300cd265ba898cdf253a30127316597796e808