Fix duplicate device sets due to path-name matching
Summary:
In Xcode 10.2, there have been some changes to the internals of Device Sets and Service Contexts that means that it's now much easier to create a device in a reference in a device set and for this to wait forever. This comes down to some assumptions that are made inside `-[SimServiceContext deviceSetWithPath:error:]`

When obtaining `SimDeviceSet` with this method, it operates on a standard multi-ton pattern. Only one Device Set per path *should* be created and the de-duplication happens as a result of string-matching (via an `NSDictionary`) on the passed in path names. If the creator of the original `SimDeviceSet` passes a path that contains symlinks, then this will misbehave:

1) `SimDeviceSet` from a path with a symlink is passed to `-[SimServiceContext deviceSetWithPath:]`, this value is cached in `-[SimServiceContext _allDeviceSets]` with the key being the path with the symlink in.
2) `-[SimDeviceSet createDeviceAsyncWithType:runtime:name:completionQueue:completionHandler:` is called on the SimDeviceSet from #1
3) Internally, this method calls `-[SimDeviceSet _awaitDeviceWithUDID:]` in a background queue. This busy-loops on `-[SimDeviceSet _devicesByUDID]`.
4) When the device is created, `-[SimServiceContext handleXPCEvent:]` is called from the `CoreSimulatorService`. It then calls `-[SimServiceContext deviceSetWithPath:]` to find the cached `SimDeviceSet`. However, this contains a path that has *resolved symlinks*.
5) In `-[SimServiceContext deviceSetWithPath:]` the `_allDeviceSets` dictionary is queried. This returns `nil` since this is string-matching of symlink vs. non-symlinked paths and therefore creates a new `SimDeviceSet` representing the same path on disk.
6) The busy-looping in `-[SimDeviceSet _awaitDeviceWithUDID:]` never ends. This is bad in the synchronous versions as the entire thread will hang forever. With `FBSimulatorControl`'s wrapping, we often apply timeouts to the async-variant.

What's interesting here is that something has changed in Xcode 10.2 in terms of how the instantiation works here, perhaps it used to resolve the symlinks before and now it expects the consumer too. Either way, this makes it possible to hang `CoreSimulator` forever by passing any path that isn't fully resolved on disk.

This means that before we construct the `SimDeviceSet` we should resolve the full path, so that it will always match the string returned in `-[SimServiceContext handleXPCEvent:]`. To make matters more interesting `-[NSString stringByResolvingSymlinksInPath]` doesn't appear to do this, but `realpath` does. This is at least the case with the symlink mapping of `/var -> /private/var` which is used a lot for temporary directory creation in tests.

Reviewed By: c-ryan747

Differential Revision: D14241937

fbshipit-source-id: 9f6998f85857a0dbdaad6915a8e8580a5821a07e