Hook CoreSimulator Logging
Summary:
It now appears that to support both ASL and `os_log`, Apple needed to add in the ability to set a custom logging function. This API is then part of the `CoreSimulator`, and we can therefore use it.

This is pretty simple, and `SimLog` is a standard format function with the function name and line number. We discard (but document) these values. We can then extract the actual format string and pass it into our logging subsystem.

Reviewed By: c-ryan747

Differential Revision: D14225899

fbshipit-source-id: 3f389617f1baf90d7fcbe4367a695e37942a73be