Ensure that cancellation is always notified in -[FBFuture onQueue:chain:]
Summary:
There was an erroneous return early in this method, which meant that consumers won't be able to re-map to some other future in the event of cancellation.

This solves two problems:
- Observers of cancellation not getting called on this method
- Re-mapping of cancellation to something else in this method.

This does mean that cancellation may be deferred, but since we need to call the `chain` block it's essentially that this deferring happens.

This was found due to the erroneous behaviour of `FBSimulatorAgentOperation`, which would not tear down the agent process upon cancellation of the public `completed` future.

Reviewed By: zbigniew-malinowski

Differential Revision: D14639114

fbshipit-source-id: 90beeffae97b70f49ef2c4e8f06c635f5d433dd1