Remove FBSimulatorPool
Summary:
This is an artefact from the days where multiple simulator sets weren't a thing and we instead used this to ensure isolation between simulators. Since this isn't a problem that `FBSimulatorControl` needs to solve, we should remove it.

Pools were only in-memory so in effect this made it harder to use, since pools would need to be allocated from the same running process. As a result this became very tricky to use. Ultimately, callers can be responsible for managing sets of simulators themselves and are encouraged to use sets to do this.

Reviewed By: zeyadsalloum, c-ryan747

Differential Revision: D14827128

fbshipit-source-id: 16e17404cac13f23e9911a0fd6dfefc438383318