Simplify data-based install
Summary:
`FBArchiveOperations` also handles zip files, which is great news for us.

There's only one place where we actually need to handle the archive-from-data, so we should localize that to the relevant place in the code.

We can also simplify the install itself, by always dealing with extracted applications. This means that we can handle the zip and tar.gz case in one.

Reviewed By: zeyadsalloum

Differential Revision: D15118604

fbshipit-source-id: 09ceadcdc3c11f38efa6c93addbe8d715d436254